fix: Update psmodulecache to v6.2.1
psmodulecache has a dependency, actions/cache. But psmodulecache v5.1 is using the recently deprecated version of actions/cache (v3.0.11).
Due to the deprecation, every ci works are failed.
Update it to the latest version which makes use of newer actions/cache

Source : https://github.blog/changelog/2024-12-05-notice-of-upcoming-releases-and-breaking-changes-for-github-actions/#actions-cache-v1-v2-and-actions-toolkit-cache-package-closing-down